1. 最小化生成，Server不会生成注释字符串 2. 优化代码生成器vargs_fmt.py 3. 修复了内存不对齐的情况下解析数据会导致部分设备错误的错误

+++ include/pcall_config.h
@@ -10,7 +10,7 @@
 /**
  * @brief PCall号的数据类型，影响PCall的最大数量。建议使用无符号数
  */
-using pcall_num_t = uint8_t;
+using pcall_num_t = uint32_t;
 
 /**
  * @brief 在服务器启用严格的长度检查
